fix remaining int to long long

+++ common.h
@@ -17,5 +17,5 @@ void fft(double* w, const long long n);
 
 void idft(double* a, const long long n);
 
-void srft(long long n, int d, int* r, int* f, int* perm, double* a, double* sa, double* dest, void (*transform) (double*, int));
+void srft(long long n, int d, int* r, int* f, int* perm, double* a, double* sa, double* dest, void (*transform) (double*, long long));
